fix: resolve all clippy warnings (12 lib + 1 bin)
- input.rs: single_match → if let for Focus::CommandPalette
- stream.rs: remove unnecessary u32 cast
- update.rs: if_not_else, useless_let_if_seq, question_mark, collapsible_if
- command_palette.rs: explicit_counter_loop → enumerate
- side_by_side.rs: single_match_else → if let (2 functions)
- text_util.rs: explicit_counter_loop → enumerate
- diff/mod.rs: elide lifetime, update caller for &mut→& change
- context.rs: needless_pass_by_ref_mut (emit_remaining_orphaned_comments)
- main.rs: allow ref_option on process_event

+++ src/stream.rs
@@ -302,7 +302,7 @@ fn comment_block_height(comments: &[Comment], content_width: u32) -> usize {
         return 0;
     }
     let max_width = content_width
-        .saturating_sub((BLOCK_SIDE_MARGIN * 2 + 1 + BLOCK_LEFT_PAD + BLOCK_RIGHT_PAD) as u32);
+        .saturating_sub(BLOCK_SIDE_MARGIN * 2 + 1 + BLOCK_LEFT_PAD + BLOCK_RIGHT_PAD);
     let max_width = max_width as usize;
     let mut content_lines = 2; // thread header line + spacing
     for comment in comments {
